6-7/integers_from_1_to_N/src/main/java/com/telran/numbers/RevealDuplicatedNumber.java
package com.telran.numbers;

import java.util.Arrays;

public class RevealDuplicatedNumber {
    public int reveal(int... numbers) {
        int length = numbers.length;

        if (length == 0)
            throw new WrongInputError();

        for (int number : numbers)
            if (number >= length || number < 0)
                throw new WrongInputError();

        Arrays.sort(numbers);

        int prev = 0;

        for (int number : numbers) {
            if (number == prev)
                return prev;
            prev = number;
        }

        return 0;
    }
}
